♻️ Modular refactor (#5)
* ♻️ Move cli logic to src/cli.js

* ♻️ Move config and particle generation into separate files

* ♻️ Move rendering logic into `utils/renderUtils.js`

* ♻️ Rename module and remove deprecated file

* 💡 Clean up comments
