+++ index.js
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+const { main } = require('./src/cli');
+
+main();
